Drawer again restructured
