+++ gui/__init__.py
@@ -1,1 +1,1 @@
-from .main_gui import MainWindow
+from .main_window import MainWindow
