+++ main.py
@@ -1,11 +1,9 @@
 import sys
-from gui import MainWindow
 from PySide6.QtWidgets import QApplication
-from gui.mechanics import Robot
+from gui.main_window import MainWindow
 
 if __name__ == "__main__":
-    app: QApplication = QApplication(sys.argv)
-    window: MainWindow = MainWindow()
+    app = QApplication(sys.argv)
+    window = MainWindow()
     window.show()
     sys.exit(app.exec())
-
